Add files via upload
Hello Guys! 
In this short coding task I have completed the third task of the data science and business analytics internship by the The Sparks Foundation. 
In this task I have performed Exploratory data analysis  on the sales data of an Online Retail Store across various states and regions of the United States and through categories and sub-categories of items.
I have done EDA as well as given some suggestions based on my opinion on various points important for the business in terms of sales and amount of profit earned.